Created Navbar container to handle login/logout switch
