+++ browser/components/navbar/NavbarContainer.js
@@ -1,0 +1,12 @@
+import React from'react';
+import { connect } from 'react-redux';
+import Navbar from './Navbar';
+import { logout } from '../../redux/user'
+
+const mapStateToProps = ({ userId }) => ({ userId });
+
+const mapDispatchToProps = dispatch => ({
+  logout: () => dispatch(logout())
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(Navbar);
